backend/requirements-local.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
# flask-migrate==4.0.5  # Removed for simplicity
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
stripe==7.8.0
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0